requirements.txt
schedule
requests
python-daemon